# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AEDC Electricity Token Subscription API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tokens:
    get:
      operationId: listTokenTransactions
      tags: [tokens]
      summary: List token purchase history
      parameters:
        - in: query
          name: meterId
          required: false
          schema:
            type: integer
        - in: query
          name: limit
          required: false
          schema:
            type: integer
        - in: query
          name: offset
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Token transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TokenTransaction"
  /tokens/purchase:
    post:
      operationId: purchaseToken
      tags: [tokens]
      summary: Purchase an electricity token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseTokenBody"
      responses:
        "201":
          description: Token purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenTransaction"
        "400":
          description: Validation or insufficient balance error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tokens/{id}:
    get:
      operationId: getTokenTransaction
      tags: [tokens]
      summary: Get a token transaction by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Token transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenTransaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Wallet ──────────────────────────────────────────────────────────────────
...
  /wallet/topup:
    post:
      operationId: topupWallet
      tags: [wallet]
      summary: Top up wallet balance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopupWalletBody"
      responses:
        "200":
          description: Wallet topped up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /wallet/transactions:
    get:
      operationId: listWalletTransactions
      tags: [wallet]
      summary: List wallet transaction history
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
        - in: query
          name: offset
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Wallet transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WalletTransaction"

  # ─── Subscriptions ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    TokenTransaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        meterId:
          type: integer
        meterNumber:
          type: string
        meterName:
          type: string
        amount:
          type: number
        token:
          type: ["string", "null"]
        units:
          type: ["number", "null"]
        status:
          type: string
        reference:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - meterId
        - meterNumber
        - meterName
        - amount
        - status
        - reference
        - createdAt
    PurchaseTokenBody:
      type: object
      properties:
        meterId:
          type: integer
        amount:
          type: number
      required:
        - meterId
        - amount

    # ─── Wallet Schemas ───────────────────────────────────────────────────────
    Wallet:
      type: object
      properties:
        balance:
          type: number
        userId:
          type: integer
      required:
        - balance
        - userId
    TopupWalletBody:
      type: object
      properties:
        amount:
          type: number
        paymentMethod:
          type: string
        reference:
          type: ["string", "null"]
      required:
        - amount
        - paymentMethod
    WalletTransaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        amount:
          type: number
        reference:
          type: string
        description:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - type
        - amount
        - reference
        - description
        - status
        - createdAt

    # ─── Subscription Schemas ─────────────────────────────────────────────────
...